+++ lib/handlers/js.js
